feat(swift): Added SwiftPackageResolvedDetectable.
